[ADD] LocationViewController (#4)
