+++ Makgulli/Presentation/LocationViewController.swift
@@ -1,0 +1,40 @@
+//
+//  LocationViewController.swift
+//  Makgulli
+//
+//  Created by 김규철 on 2023/09/25.
+//
+
+import UIKit
+import NMapsMap
+
+import RxSwift
+import RxRelay
+import RxCocoa
+
+final class LocationViewController: UIViewController {
+    private let viewModel = LocationViewModel(searchLocationUseCase: DefaultSearchLocationUseCase(searchLocationRepository: DefaultSearchLocationRepository(networkManager: NetworkManager())))
+    
+    var bag: DisposeBag = .init()
+        
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        self.view.backgroundColor = .systemPink
+        
+        let mapView = NMFMapView(frame: view.frame)
+        view.addSubview(mapView)
+        bind()
+        
+    }
+    
+    private func bind() {
+        let input = LocationViewModel.Input(viewDidLoadEvent: Observable.just(()).asObservable())
+        let output = viewModel.transform(input: input)
+        
+        output.locationVO
+            .bind(onNext: { searchLocationVO in
+                print(searchLocationVO)
+            })
+            .disposed(by: bag)
+    }
+}
